+++ src/main/kotlin/no/nav/familie/ef/sak/api/dto/BehandlingHistorikkDto.kt
@@ -4,7 +4,7 @@ import no.nav.familie.ef.sak.service.steg.StegType
 import java.time.LocalDateTime
 import java.util.*
 
-data class BehandlingsHistorikkDto(val behandlingId: UUID,
+data class BehandlingshistorikkDto(val behandlingId: UUID,
                                    var steg: StegType,
                                    val endretAvNavn: String,
                                    val endretAvMail: String,
