+++ src/main/kotlin/no/nav/familie/ef/sak/repository/BehandlingHistorikkRepository.kt
@@ -1,12 +1,12 @@
 package no.nav.familie.ef.sak.repository
 
-import no.nav.familie.ef.sak.repository.domain.BehandlingHistorikk
+import no.nav.familie.ef.sak.repository.domain.Behandlingshistorikk
 import org.springframework.stereotype.Repository
 import java.util.*
 
 @Repository
-interface BehandlingHistorikkRepository : RepositoryInterface<BehandlingHistorikk, UUID>,
-                                          InsertUpdateRepository<BehandlingHistorikk> {
+interface BehandlingHistorikkRepository : RepositoryInterface<Behandlingshistorikk, UUID>,
+                                          InsertUpdateRepository<Behandlingshistorikk> {
 
-    fun findByBehandlingId(behandlingId: UUID): List<BehandlingHistorikk>
+    fun findByBehandlingId(behandlingId: UUID): List<Behandlingshistorikk>
 }
