+++ src/main/kotlin/no/nav/familie/ef/sak/service/BehandlingHistorikkService.kt
@@ -1,19 +1,18 @@
 package no.nav.familie.ef.sak.service
 
-import no.nav.familie.ef.sak.api.dto.BehandlingsHistorikkDto
 import no.nav.familie.ef.sak.repository.BehandlingHistorikkRepository
-import no.nav.familie.ef.sak.repository.domain.BehandlingHistorikk
+import no.nav.familie.ef.sak.repository.domain.Behandlingshistorikk
 import org.springframework.stereotype.Service
 import java.util.*
 
 @Service
 class BehandlingHistorikkService(val behandlingHistorikkRepository: BehandlingHistorikkRepository) {
 
-    fun finnBehandlingHistorikk(behandlingId: UUID) : List<BehandlingHistorikk> {
+    fun finnBehandlingHistorikk(behandlingId: UUID) : List<Behandlingshistorikk> {
         return behandlingHistorikkRepository.findByBehandlingId(behandlingId)
     }
 
-    fun opprettHistorikkInnslag(behandlingHistorikk: BehandlingHistorikk) {
-        behandlingHistorikkRepository.insert(behandlingHistorikk)
+    fun opprettHistorikkInnslag(behandlingshistorikk: Behandlingshistorikk) {
+        behandlingHistorikkRepository.insert(behandlingshistorikk)
     }
 }
