+++ src/main/kotlin/no/nav/familie/ef/sak/service/steg/StegService.kt
@@ -4,7 +4,7 @@ import io.micrometer.core.instrument.Counter
 import io.micrometer.core.instrument.Metrics
 import no.nav.familie.ef.sak.config.RolleConfig
 import no.nav.familie.ef.sak.repository.domain.Behandling
-import no.nav.familie.ef.sak.repository.domain.BehandlingHistorikk
+import no.nav.familie.ef.sak.repository.domain.Behandlingshistorikk
 import no.nav.familie.ef.sak.service.BehandlingHistorikkService
 import no.nav.familie.ef.sak.service.BehandlingService
 import no.nav.familie.ef.sak.service.steg.StegType.BEHANDLING_FERDIGSTILT
@@ -70,10 +70,10 @@ class StegService(private val behandlingSteg: List<BehandlingSteg<*>>,
 
             behandlingSteg.validerSteg(behandling)
 
-            behandlingHistorikkService.opprettHistorikkInnslag(BehandlingHistorikk(behandlingId = behandling.id,
-                                                                                   steg = behandling.steg,
-                                                                                   endretAvNavn = saksbehandlerNavn,
-                                                                                   endretAvMail = SikkerhetContext.hentSaksbehandler()))
+            behandlingHistorikkService.opprettHistorikkInnslag(Behandlingshistorikk(behandlingId = behandling.id,
+                                                                                    steg = behandling.steg,
+                                                                                    endretAvNavn = saksbehandlerNavn,
+                                                                                    endretAvMail = SikkerhetContext.hentSaksbehandler()))
 
             val nesteSteg = behandlingSteg.utførOgReturnerNesteSteg(behandling, data)
