+++ src/test/kotlin/no/nav/familie/ef/sak/OppslagSpringRunnerTest.kt
@@ -65,7 +65,7 @@ abstract class OppslagSpringRunnerTest {
                TilkjentYtelse::class,
                Oppgave::class,
                Vilkårsvurdering::class,
-               BehandlingHistorikk::class,
+               Behandlingshistorikk::class,
                Behandling::class,
                Fagsak::class,
         ).forEach { jdbcAggregateOperations.deleteAll(it.java) }
