+++ src/test/kotlin/no/nav/familie/ef/sak/service/BehandlingshistorikkServiceTest.kt
@@ -6,7 +6,7 @@ import no.nav.familie.ef.sak.no.nav.familie.ef.sak.repository.fagsak
 import no.nav.familie.ef.sak.repository.BehandlingHistorikkRepository
 import no.nav.familie.ef.sak.repository.BehandlingRepository
 import no.nav.familie.ef.sak.repository.FagsakRepository
-import no.nav.familie.ef.sak.repository.domain.BehandlingHistorikk
+import no.nav.familie.ef.sak.repository.domain.Behandlingshistorikk
 import no.nav.familie.ef.sak.service.BehandlingHistorikkService
 import no.nav.familie.ef.sak.sikkerhet.SikkerhetContext
 import org.assertj.core.api.Assertions.assertThat
@@ -15,7 +15,7 @@ import org.springframework.beans.factory.annotation.Autowired
 import java.time.LocalDateTime
 
 
-internal class BehandlingHistorikkServiceTest : OppslagSpringRunnerTest() {
+internal class BehandlingshistorikkServiceTest : OppslagSpringRunnerTest() {
 
     @Autowired private lateinit var behandlingHistorikkService: BehandlingHistorikkService
     @Autowired private lateinit var behandlingRepository: BehandlingRepository
@@ -28,14 +28,14 @@ internal class BehandlingHistorikkServiceTest : OppslagSpringRunnerTest() {
         /** Lagre */
         val fagsak = fagsakRepository.insert(fagsak())
         val behandling = behandlingRepository.insert(behandling(fagsak))
-        val behandlingHistorikk = behandlingHistorikkRepository.insert(BehandlingHistorikk(behandlingId = behandling.id,
-                                                                                           steg = behandling.steg,
-                                                                                           endretAvNavn = "Saksbehandlernavn",
-                                                                                           endretAvMail = SikkerhetContext.hentSaksbehandler(),
-                                                                                           endretTid = LocalDateTime.now()))
+        val behandlingHistorikk = behandlingHistorikkRepository.insert(Behandlingshistorikk(behandlingId = behandling.id,
+                                                                                            steg = behandling.steg,
+                                                                                            endretAvNavn = "Saksbehandlernavn",
+                                                                                            endretAvMail = SikkerhetContext.hentSaksbehandler(),
+                                                                                            endretTid = LocalDateTime.now()))
 
         /** Hent */
-        val innslag : BehandlingHistorikk = behandlingHistorikkService.finnBehandlingHistorikk(behandling.id)[0]
+        val innslag : Behandlingshistorikk = behandlingHistorikkService.finnBehandlingHistorikk(behandling.id)[0]
 
         assertThat(innslag).isEqualTo(behandlingHistorikk)
     }
